Refactor directories within routes to ensure that end of service report pages are contained in directories whose name matches the route
