+++ server/routes/probation-practitioner/end-of-service-report/endOfServiceReportPresenter.ts
@@ -1,8 +1,8 @@
-import EndOfServiceReport from '../../models/endOfServiceReport'
-import SentReferral from '../../models/sentReferral'
-import ServiceCategory from '../../models/serviceCategory'
-import EndOfServiceReportAnswersPresenter from '../shared/endOfServiceReportAnswersPresenter'
-import PresenterUtils from '../../utils/presenterUtils'
+import EndOfServiceReport from '../../../models/endOfServiceReport'
+import SentReferral from '../../../models/sentReferral'
+import ServiceCategory from '../../../models/serviceCategory'
+import EndOfServiceReportAnswersPresenter from '../../shared/endOfServiceReportAnswersPresenter'
+import PresenterUtils from '../../../utils/presenterUtils'
 
 export default class EndOfServiceReportPresenter {
   constructor(
