+++ server/routes/probationPractitionerReferrals/probationPractitionerReferralsController.ts
@@ -13,8 +13,8 @@ import SubmittedPostSessionFeedbackPresenter from '../shared/action-plan/appoint
 import SubmittedPostSessionFeedbackView from '../shared/action-plan/appointment/post-session-feedback/submittedPostSessionFeedbackView'
 import ReferralCancellationReasonPresenter from './referralCancellationReasonPresenter'
 import ReferralCancellationReasonView from './referralCancellationReasonView'
-import EndOfServiceReportPresenter from './endOfServiceReportPresenter'
-import EndOfServiceReportView from './endOfServiceReportView'
+import EndOfServiceReportPresenter from '../probation-practitioner/end-of-service-report/endOfServiceReportPresenter'
+import EndOfServiceReportView from '../probation-practitioner/end-of-service-report/endOfServiceReportView'
 import ReferralCancellationCheckAnswersPresenter from './referralCancellationCheckAnswersPresenter'
 import ReferralCancellationCheckAnswersView from './referralCancellationCheckAnswersView'
 import { FormValidationError } from '../../utils/formValidationError'
