+++ server/routes/service-provider/end-of-service-report/check-answers/endOfServiceReportCheckAnswersPresenter.test.ts
@@ -1,7 +1,7 @@
 import EndOfServiceReportCheckAnswersPresenter from './endOfServiceReportCheckAnswersPresenter'
-import sentReferralFactory from '../../../testutils/factories/sentReferral'
-import serviceCategoryFactory from '../../../testutils/factories/serviceCategory'
-import endOfServiceReportFactory from '../../../testutils/factories/endOfServiceReport'
+import sentReferralFactory from '../../../../../testutils/factories/sentReferral'
+import serviceCategoryFactory from '../../../../../testutils/factories/serviceCategory'
+import endOfServiceReportFactory from '../../../../../testutils/factories/endOfServiceReport'
 
 describe(EndOfServiceReportCheckAnswersPresenter, () => {
   const serviceCategory = serviceCategoryFactory.build({
