+++ server/routes/service-provider/end-of-service-report/confirmation/endOfServiceReportConfirmationPresenter.ts
@@ -1,8 +1,8 @@
-import SentReferral from '../../models/sentReferral'
-import ServiceCategory from '../../models/serviceCategory'
-import PresenterUtils from '../../utils/presenterUtils'
-import { SummaryListItem } from '../../utils/summaryList'
-import utils from '../../utils/utils'
+import SentReferral from '../../../../models/sentReferral'
+import ServiceCategory from '../../../../models/serviceCategory'
+import PresenterUtils from '../../../../utils/presenterUtils'
+import { SummaryListItem } from '../../../../utils/summaryList'
+import utils from '../../../../utils/utils'
 
 export default class EndOfServiceReportConfirmationPresenter {
   constructor(private readonly referral: SentReferral, private readonly serviceCategory: ServiceCategory) {}
