+++ server/routes/service-provider/end-of-service-report/further-information/endOfServiceReportFurtherInformationForm.test.ts
@@ -1,4 +1,4 @@
-import TestUtils from '../../../testutils/testUtils'
+import TestUtils from '../../../../../testutils/testUtils'
 import EndOfServiceReportFurtherInformationForm from './endOfServiceReportFurtherInformationForm'
 
 describe(EndOfServiceReportFurtherInformationForm, () => {
