+++ server/routes/service-provider/end-of-service-report/further-information/endOfServiceReportFurtherInformationForm.ts
@@ -1,5 +1,5 @@
 import { Request } from 'express'
-import { UpdateDraftEndOfServiceReportParams } from '../../services/interventionsService'
+import { UpdateDraftEndOfServiceReportParams } from '../../../../services/interventionsService'
 
 export default class EndOfServiceReportFurtherInformationForm {
   constructor(private readonly request: Request) {}
