+++ server/routes/service-provider/end-of-service-report/outcomes/endOfServiceReportOutcomeForm.ts
@@ -1,10 +1,10 @@
 import * as ExpressValidator from 'express-validator'
 import { Request } from 'express'
-import { FormData } from '../../utils/forms/formData'
-import errorMessages from '../../utils/errorMessages'
-import { UpdateDraftEndOfServiceReportParams } from '../../services/interventionsService'
-import ServiceUser from '../../models/serviceUser'
-import FormUtils from '../../utils/formUtils'
+import { FormData } from '../../../../utils/forms/formData'
+import errorMessages from '../../../../utils/errorMessages'
+import { UpdateDraftEndOfServiceReportParams } from '../../../../services/interventionsService'
+import ServiceUser from '../../../../models/serviceUser'
+import FormUtils from '../../../../utils/formUtils'
 
 export default class EndOfServiceReportOutcomeForm {
   constructor(
