+++ server/routes/service-provider/end-of-service-report/outcomes/endOfServiceReportOutcomePresenter.test.ts
@@ -1,8 +1,8 @@
 import EndOfServiceReportOutcomePresenter from './endOfServiceReportOutcomePresenter'
-import sentReferralFactory from '../../../testutils/factories/sentReferral'
-import endOfServiceReportFactory from '../../../testutils/factories/endOfServiceReport'
-import serviceCategoryFactory from '../../../testutils/factories/serviceCategory'
-import { EndOfServiceReportOutcome } from '../../models/endOfServiceReport'
+import sentReferralFactory from '../../../../../testutils/factories/sentReferral'
+import endOfServiceReportFactory from '../../../../../testutils/factories/endOfServiceReport'
+import serviceCategoryFactory from '../../../../../testutils/factories/serviceCategory'
+import { EndOfServiceReportOutcome } from '../../../../models/endOfServiceReport'
 
 describe(EndOfServiceReportOutcomePresenter, () => {
   const serviceCategory = serviceCategoryFactory.build({ name: 'social inclusion' })
