+++ server/routes/serviceProviderReferrals/serviceProviderReferralsController.ts
@@ -43,16 +43,16 @@ import PostSessionFeedbackCheckAnswersView from '../service-provider/action-plan
 import PostSessionFeedbackCheckAnswersPresenter from '../service-provider/action-plan/appointment/post-session-feedback/check-your-answers/postSessionFeedbackCheckAnswersPresenter'
 import SubmittedPostSessionFeedbackView from '../shared/action-plan/appointment/post-session-feedback/submittedPostSessionFeedbackView'
 import SubmittedPostSessionFeedbackPresenter from '../shared/action-plan/appointment/post-session-feedback/submittedPostSessionFeedbackPresenter'
-import EndOfServiceReportOutcomeForm from './endOfServiceReportOutcomeForm'
-import EndOfServiceReportOutcomePresenter from './endOfServiceReportOutcomePresenter'
-import EndOfServiceReportOutcomeView from './endOfServiceReportOutcomeView'
-import EndOfServiceReportFurtherInformationForm from './endOfServiceReportFurtherInformationForm'
-import EndOfServiceReportFurtherInformationPresenter from './endOfServiceReportFurtherInformationPresenter'
-import EndOfServiceReportFurtherInformationView from './endOfServiceReportFurtherInformationView'
-import EndOfServiceReportCheckAnswersPresenter from './endOfServiceReportCheckAnswersPresenter'
-import EndOfServiceReportCheckAnswersView from './endOfServiceReportCheckAnswersView'
-import EndOfServiceReportConfirmationPresenter from './endOfServiceReportConfirmationPresenter'
-import EndOfServiceReportConfirmationView from './endOfServiceReportConfirmationView'
+import EndOfServiceReportOutcomeForm from '../service-provider/end-of-service-report/outcomes/endOfServiceReportOutcomeForm'
+import EndOfServiceReportOutcomePresenter from '../service-provider/end-of-service-report/outcomes/endOfServiceReportOutcomePresenter'
+import EndOfServiceReportOutcomeView from '../service-provider/end-of-service-report/outcomes/endOfServiceReportOutcomeView'
+import EndOfServiceReportFurtherInformationForm from '../service-provider/end-of-service-report/further-information/endOfServiceReportFurtherInformationForm'
+import EndOfServiceReportFurtherInformationPresenter from '../service-provider/end-of-service-report/further-information/endOfServiceReportFurtherInformationPresenter'
+import EndOfServiceReportFurtherInformationView from '../service-provider/end-of-service-report/further-information/endOfServiceReportFurtherInformationView'
+import EndOfServiceReportCheckAnswersPresenter from '../service-provider/end-of-service-report/check-answers/endOfServiceReportCheckAnswersPresenter'
+import EndOfServiceReportCheckAnswersView from '../service-provider/end-of-service-report/check-answers/endOfServiceReportCheckAnswersView'
+import EndOfServiceReportConfirmationPresenter from '../service-provider/end-of-service-report/confirmation/endOfServiceReportConfirmationPresenter'
+import EndOfServiceReportConfirmationView from '../service-provider/end-of-service-report/confirmation/endOfServiceReportConfirmationView'
 import ControllerUtils from '../../utils/controllerUtils'
 import AuthUserDetails from '../../models/hmppsAuth/authUserDetails'
 import ServiceCategory from '../../models/serviceCategory'
